refactor(agents): split CLI runner concepts
